requirements.txt
asgiref==3.8.1
black==24.3.0
blinker==1.7.0
build==1.1.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
identify==2.5.35
idna==3.6
importlib_metadata==7.1.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
keyring==24.3.1
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
pre-commit==3.7.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.6.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
trove-classifiers==2024.3.3
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
xattr==1.1.0
zipp==3.18.1
